refactor: apply suggestions from code review & update importations & lint
