+++ index.ts
@@ -1,11 +1,11 @@
-import PalavrecoClient from "./src/PalavrecoClient";
-import { PalavrecoConfig, PalavrecoOptions } from "./src/utils/PalavrecoConfig";
+import PalavrecoClient from './src/PalavrecoClient';
+import { PalavrecoConfig, PalavrecoOptions } from './src/utils/PalavrecoConfig';
 
 const client = new PalavrecoClient({
-    intents: PalavrecoConfig.intents,
+	intents: PalavrecoConfig.intents,
 });
 
 client.start({
-    commands: client.commands,
-    token: PalavrecoOptions.token,
+	commands: client.commands,
+	token: PalavrecoOptions.token,
 });
