+++ src/utils/PalavrecoConfig.ts
@@ -1,31 +1,29 @@
-import { ActivityType, Collection, IntentsString, PermissionString, Snowflake } from "discord.js";
-import { PalavrecoCommands } from "../scructures/BaseCommand";
+import { Collection } from 'discord.js';
+import { PalavrecoCommand } from '../structures/BaseCommand';
 
-export const PalavrecoConfig = {
-    developers: [
-        '392099429768888351', '535549399938367528',
-        '762079458249146388'
-    ] as Snowflake[],
-    embedColor: '#2596be' as string,
-    intents: ['GUILDS', 'GUILD_MESSAGES'] as IntentsString[],
-    defaultPermissions: [
-        'SEND_MESSAGES', 'MANAGE_MESSAGES',
-        'USE_EXTERNAL_EMOJIS', 'VIEW_CHANNEL'
-    ] as PermissionString[],
-    activities: [
-        { name: '/adivinhar', type: 'PLAYING' },
-        { name: 'suas tentativas...', type: 'WATCHING' },
-    ] as PalavrecoActivity[],
-    supportServer: 'https://discord.gg/KEdytHNbK2' as string,
-    botInvite: 'https://discord.com/oauth2/authorize?client_id=935291567038672906&permissions=273408&scope=bot%20applications.commands' as string,
+import { PalavrecoFundamentals, PalavrecoSettings } from '../structures/ClientSettings';
+export const PalavrecoConfig: PalavrecoSettings = {
+	developers: [
+		'392099429768888351', '535549399938367528',
+		'762079458249146388',
+	],
+	embedColor: '#2596be',
+	intents: ['GUILDS', 'GUILD_MESSAGES'],
+	defaultPermissions: [
+		'SEND_MESSAGES', 'MANAGE_MESSAGES',
+		'USE_EXTERNAL_EMOJIS', 'VIEW_CHANNEL',
+	],
+	activities: [
+		{ name: '/adivinhar', type: 'PLAYING' },
+		{ name: 'suas tentativas...', type: 'WATCHING' },
+	],
+	supportServer: 'https://discord.gg/KEdytHNbK2',
+	// eslint-disable-next-line max-len
+	botInvite: 'https://discord.com/oauth2/authorize?client_id=935291567038672906&permissions=273408&scope=bot%20applications.commands',
 };
 
-export interface PalavrecoActivity {
-    name: string;
-    type: ActivityType;
-}
-
-export const PalavrecoOptions = {
-    commands: new Collection<string, PalavrecoCommands>(),
-    token: process.env.TOKEN as string,
-}
+export const PalavrecoOptions: PalavrecoFundamentals = {
+	commands: new Collection<string, PalavrecoCommand>(),
+	// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+	token: process.env.TOKEN!,
+};
